some code refactoring. update readme
