+++ bot/crm_mock/db_utils.py
@@ -1,3 +1,5 @@
+""""""
+
 from os import getenv
 
 import aiosqlite
