--- bot/test.json
@@ -1,48 +1,0 @@
-{
-    anser: [
-        (
-            AgentAction(
-                tool="Use the ner_booking_tool to check availability.",
-                tool_input="Can I have my nails done 26th of september at 10am",
-                log='The user wants to know if they can have their nails done on a specific date and time.\nAction: Use the ner_booking_tool to check availability.\nAction Input: "Can I have my nails done 26th of september at 10am"',
-            ),
-            "Use the ner_booking_tool to check availability. is not a valid tool, try one of [sql_tool, ner_booking_tool].",
-        ),
-        (
-            AgentAction(
-                tool="_Exception",
-                tool_input="Invalid Format: Missing 'Action:' after 'Thought:",
-                log="I need to use the ner_booking_tool to check availability.",
-            ),
-            "Invalid Format: Missing 'Action:' after 'Thought:",
-        ),
-        (
-            AgentAction(
-                tool="ner_booking_tool",
-                tool_input="Can I have my nails done 26th of september at 10am",
-                log='I need to use the ner_booking_tool to check availability.\nAction: ner_booking_tool\nAction Input: "Can I have my nails done 26th of september at 10am"',
-            ),
-            {
-                "action": "Create",
-                "master_name": None,
-                "beauty_service": "Manicure",
-                "booking_date": "2023-09-26",
-                "booking_time": "10:00",
-            },
-        ),
-        (
-            AgentAction(
-                tool="ner_booking_tool",
-                tool_input="{'action': 'Create', 'master_name': None, 'beauty_service': 'Manicure', 'booking_date': '2023-09-26', 'booking_time': '10:00'}",
-                log="The user wants to create a booking for a manicure on September 26th at 10am.\nAction: ner_booking_tool\nAction Input: {'action': 'Create', 'master_name': None, 'beauty_service': 'Manicure', 'booking_date': '2023-09-26', 'booking_time': '10:00'}",
-            ),
-            {
-                "action": "Create",
-                "master_name": None,
-                "beauty_service": "Manicure",
-                "booking_date": "2023-09-26",
-                "booking_time": "10:00",
-            },
-        ),
-    ]
-}
